% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{class}
\name{ipsAdaptive}
\alias{ipsAdaptive}
\title{ipsAdaptive}
\description{
ipsAdaptive is an S4 class object. The object is the output of
the \code{rank.compare} or \code{findIPs} function, when "adaptive" is selected
for method.
}
\section{Slots}{

\describe{
\item{\code{kappa}}{a numeric. The shape parameter for the weight function.}

\item{\code{score}}{a numeric vector. The influential scores.}

\item{\code{origRank}}{a character vector. The original ranking.}

\item{\code{drop1Rank}}{a matrix. Leave-one-out rankings.}

\item{\code{origRankWeighted}}{a numeric vector. The weighted original ranking.}

\item{\code{drop1RankWeighted}}{a matrix. Weighted leave-one-out rankings.}
}}

