% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\docType{class}
\name{ipsOthers}
\alias{ipsOthers}
\title{ipsOthers}
\value{
S4 class object.
}
\description{
ipsOthers is an S4 class object. The object is the output of
the \code{rank.compare} or \code{findIPs} function, when "weightedSpearman"
or "unweighted" is selected.
}
\section{Slots}{

\describe{
\item{\code{score}}{a numeric vector. The influential scores.}

\item{\code{origRank}}{a character vector. The original ranking.}

\item{\code{drop1Rank}}{a matrix. Leave-one-out rankings.}
}}

