% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miller05.R
\docType{data}
\name{miller05}
\alias{miller05}
\title{miller05 data}
\format{
a list
}
\usage{
data(miller05)
}
\description{
miller05 is gene expression data with 1000 genes randomly sampled from 22283 genes
and 236 samples since removing the case with missing response.
The data has binary and survival response. The binary response
contains 58 case with p53 mutant and 193 wild type mutant. The survival
response has a total of 55 events.
}
\examples{
data(miller05)
}
\references{
Miller, Lance D., et al.
"An expression signature for p53 status in human breast cancer predicts
mutation status, transcriptional effects, and patient survival."
Proceedings of the National Academy of Sciences 102.38 (2005):
13550-13555.doi:10.1073pnas.0506230102
}
\keyword{datasets}
