% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,ipsAdaptive,missing-method}
\alias{plot,ipsAdaptive,missing-method}
\title{Visualize the influential scores}
\usage{
\S4method{plot}{ipsAdaptive,missing}(x, topn = 5, ylim = NULL, ...)
}
\arguments{
\item{x}{the object obtained from \code{rank.compare} or \code{findIPs} function}

\item{topn}{the top n most influential points to be labelled in the plot.}

\item{ylim}{y coordinates ranges}

\item{...}{other arguments}
}
\description{
Visualize influential score using lollipop plot. The function is an S4 class
method for object obtained from \code{rank.compare} or \code{findIPs} function.
}
\examples{

data(miller05)
X <- miller05$X
y <- miller05$y
obj <- getdrop1ranks(X, y,
                     fun = 't.test',
                     decreasing = FALSE,
                     topN = 100)
rks <- sumRanks(origRank = obj$origRank,
                drop1Rank = obj$drop1Rank,
                topN = 100,
                method = 'adaptive')
plot(rks, topn = 5, ylim = NULL)

}
