% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,ipsOthers,missing-method}
\alias{plot,ipsOthers,missing-method}
\title{Plot method for S4 class ipsOthers}
\usage{
\S4method{plot}{ipsOthers,missing}(x, topn = 5, ylim = NULL, ...)
}
\arguments{
\item{x}{An object of S4 class ipsOthers.}

\item{topn}{the top n most influential points to be labelled in the plot.}

\item{ylim}{y coordinates ranges}

\item{...}{Additional parameters.}
}
\description{
Plot method for S4 class ipsOthers.
}
