% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ips.R
\name{plot.ips}
\alias{plot.ips}
\title{Visualize the influential scores}
\usage{
\method{plot}{ips}(obj, topn = 5, ylim = NULL)
}
\arguments{
\item{obj}{the objective obtained from \code{rank.compare} or \code{ipsr} function}

\item{topn}{the top n most influential points to be labelled in the plot.}

\item{ylim}{y coordinates ranges}
}
\description{
Visualize influential score using lollipop plot
}
\examples{

data(miller05)
X <- miller05$X
y <- miller05$y
obj <- getdrop1ranks(X, y,
                     fun = "t.test",
                     decreasing = FALSE,
                     topN = 100)
rks <- sumRanks(origRank = obj$origRank,
                drop1Rank = obj$drop1Rank,
                topN = 100,
                method = "adaptive")
plot(rks, topn = 5, ylim = NULL)

}
