% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsr.R
\name{plot_unweighted_ranks}
\alias{plot_unweighted_ranks}
\title{Visualize the unweighted rank changes}
\usage{
plot_unweighted_ranks(obj, top = TRUE, points.arg = list(), top.arg = list())
}
\arguments{
\item{obj}{the objective obtained from rank.compare function}

\item{top}{logical, whether the most influential case needs to be plot}

\item{points.arg}{arguments pass to points}

\item{top.arg}{arguments pass to points of top feature}
}
\description{
Compare the unweighted rank changes. The plot displays the original ranking
leave-1-out rankings.
}
\examples{
X <- miller05$X
y <- miller05$y

obj <- getdrop1ranks(X, y, topn = 100, fun = "t.test")
rks <-rank.compare(refr = obj$orig, matr = obj$drop1rank, topn = 100, dist = "L2")
par(mfrow = c(1, 2), mar = c(4, 4, 2, 2))
plot_unweighted_ranks(rks, top = TRUE)
plot_kappa_weight(rks$kappa, n = 100, type = "line")
plot(rks, topn = 5, ylim = NULL)
}
\author{
Shuo Wang
}
