% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsr.R
\name{rank.compare}
\alias{rank.compare}
\title{Measure the influence of the cases on feature rankings}
\usage{
rank.compare(refr, matr, topn = NULL, dist = c("L1", "L2"), ...)
}
\arguments{
\item{refr}{vectors, reference rankings.For influential observation detection,
refr is the original ranking obtained with the whole data.}

\item{matr}{matrix or data.frame, Each column is a feature list with a case
removed.}

\item{topn}{the top n features in refr will be used for rank comparison.
If null, include all features.}

\item{dist}{the strategy to calculate the rank changes. Choose between L1 and L2 norm.}

\item{...}{other arguments}
}
\description{
This function measures the rank changes due to case deletion. The process incorporates
a rank comparison method taking the feature importance into account. The weight function
is determined by the distribution of rank changes.
}
\examples{
 
X <- miller05$X
y <- miller05$y

obj <- getdrop1ranks(X, y, topn = 100, fun = "t.test")
rks <-rank.compare(refr = obj$orig, matr = obj$drop1rank, topn = 100, dist = "L2")
par(mfrow = c(1, 2), mar = c(4, 4, 2, 2))
plot_unweighted_ranks(rks, top = TRUE)
plot_kappa_weight(rks$kappa, n = 100, type = "line")
plot(rks, topn = 5, ylim = NULL)
}
\author{
Shuo Wang
}
