% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsr.R
\name{getdrop1ranks}
\alias{getdrop1ranks}
\title{Derive ranking lists including original and leave-1-out rankings}
\usage{
getdrop1ranks(
  X,
  y,
  fun = c("t.test", "cox"),
  decreasing = FALSE,
  topn = 100,
  ncores = NULL
)
}
\arguments{
\item{X}{A data matrix, with columns being the variables and rows being samples}

\item{y}{Groups or survival object (for cox regression)}

\item{fun}{function to obtain ranking criteria with x and y being the input, where
x a variables in X, y is the response. "t.test" and "cox" are two simple options
that rank features according to the p.values of t.test or univariate cox regression.}

\item{decreasing}{logical. How the rank criteria are ordered? For instance,
p-value should be ordered increasingly, while fold-change should be ordered decreasingly.}

\item{topn}{the number of important features included for comparison.
The top n features in the original ranking list.}

\item{ncores}{the number of CPU cores used for parallel running.
If ncores = NULL, a single core is used.}
}
\description{
This function calculate the feature rankings and the features rankings with
one-case deletion with a given ranking criteria function.
}
\examples{
X <- miller05$X
y <- miller05$y

obj <- getdrop1ranks(X, y, topn = 100, fun = "t.test")
rks <- rank.compare(refr = obj$orig, matr = obj$drop1rank, topn = 100, dist = "L2")
par(mfrow = c(1, 2), mar = c(4, 4, 2, 2))
plot_unweighted_ranks(rks, top = TRUE)
plot_kappa_weight(rks$kappa, n = 100, type = "line")
plot(rks, topn = 5, ylim = NULL)

}
\author{
Shuo Wang
}
